% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_dat}
\alias{full_dat}
\title{mHealth Data}
\format{
## `full_dat`
A data frame with 252 rows and 13 columns:
\describe{
  \item{date}{Date of collection}
  \item{subject_id}{Participant ID, either 1 or 2}
  \item{filtered_step_count, filtered_time_at_home, filtered_conversation_percent,
   filtered_tic_voiced_time, filtered_sleep_duration, filtered_travel_diameter,
   filtered_active_time, filtered_total_activity_duration,
   filtered_total_location_duration, filtered_sleep_interruptions,
   filtered_radius_of_gyration}{mHealth data collected}
}
}
\source{
Weill Cornell Advanced Laboratory for Accelerating the Reach and Impact of Treatment for Mid- and Late-Life Depression (ALACRITY) Research Center
data(full_dat)
}
\usage{
full_dat
}
\description{
A subset of mHealth data collected from elderly patients with major depressive
disorder by Weill Cornell Advanced Laboratory for Accelerating the Reach and Impact of
Treatment for Mid- and Late-Life Depression (ALACRITY) Research Center
}
\keyword{datasets}
