% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSpamH.R
\name{TwoSpamH}
\alias{TwoSpamH}
\title{The 2SpamH algorithm}
\usage{
TwoSpamH(
  data,
  variable,
  PC.vars,
  step2.var = c(""),
  imp.method = na_mean,
  thresholds = list(low = c(0.3), high = c(0.7)),
  num.neighbor = 5,
  check.cor = 0.8,
  plot.data = F,
  seed = NULL
)
}
\arguments{
\item{data}{A data frame contains the variables to be labelled and the variables on which principle component analysis is performed (PC vars).}

\item{variable}{A variable name or the position indexes of the variable to be labelled.}

\item{PC.vars}{A list object, such that each element of this list contains the vectors of either the name of the PC vars or the position indexes of the PC vars in the data frame.}

\item{step2.var}{A vector of variable names to be added into the step 2 KNN feature space.}

\item{imp.method}{A function which serves as the imputation method for missing data in PC vars. This function should take a vector with missing and retun a vector without.}

\item{thresholds}{A list of which the first element contains the 'low' quantile thresholds for each group of PC var, and the second for 'high'. Each element in it should be a vector eitehr of the same length as the number of PC var groups or 1.}

\item{num.neighbor}{Number of the neibors considered by each unlabelled data points in stage 2}

\item{check.cor}{Whether the highly correlated variables should be removed when performing stage 2 of the TSknn. If no, input should be NULL. If yes, input shou be the correlation threshold for variables to be removed.}

\item{plot.data}{If TRUE, the outputted results are for plotting. If FALSE, the function outputs the original data frame where the filtered variable is labelled with extra NAs.}

\item{seed}{The seed to be set, default is NULL.}
}
\value{
Either a dataframe object with filtered \code{variable} or a list for visualization of the labelling algorithm.
}
\description{
This function performs the 2SpamH algorithm on a ceratin variable in the input data
}
